\name{stmgp-package}
\alias{stmgp-package}
\docType{package}
\title{
\packageTitle{stmgp}
}
\description{
\packageDescription{stmgp}
}
\details{

The DESCRIPTION file:
\packageIndices{stmgp}
}
\author{
Maintainer: \packageMaintainer{stmgp}
}
\references{
Ueki M, Tamiya G, for Alzheimer's Disease Neuroimaging Initiative (2016) Smooth-thresholdmultivariate genetic prediction with unbiased model selection. Genet Epidemiol 40:233-43. <https://doi.org/10.1002/gepi.21958>

Ueki M (2009) A note on automatic variable selection using smooth-threshold estimating equations. Biometrika 96:1005-11. <https://doi.org/10.1093/biomet/asp060>

Ueki M, Fujii M, Tamiya G, for Alzheimer's Disease Neuroimaging Initiative and the Alzheimer's Disease Metabolomics Consortium (2019) Quick assessment for systematic test statistic inflation/deflation due to null model misspecifications in genome-wide environment interaction studies. PLoS ONE 14:e0219825. <https://doi.org/10.1371/journal.pone.0219825>

Ueki M, Tamiya G, for Alzheimer's Disease Neuroimaging Initiative (2021) mooth-threshold multivariate genetic prediction incorporating gene-nvironment interactions. G3 Genes|Genomes|Genetics 11:jkab278. <https://doi.org/10.1093/g3journal/jkab278>


}
\keyword{ package }

