\name{stmgp-package}
\alias{stmgp-package}
\docType{package}
\title{
\packageTitle{stmgp}
}
\description{
\packageDescription{stmgp}
}
\details{

The DESCRIPTION file:
\packageIndices{stmgp}
}
\author{
Maintainer: \packageMaintainer{stmgp}
}
\references{
Ueki M, Tamiya G, and for Alzheimer's Disease Neuroimaging Initiative. (2016)
Smooth-threshold multivariate genetic prediction with unbiased model selection.
Genet Epidemiol 40:233-43.
}
\keyword{ package }

