\name{stmgp-package}
\alias{stmgp-package}
\docType{package}
\title{
\packageTitle{stmgp}
}
\description{
\packageDescription{stmgp}
}
\details{

The DESCRIPTION file:
\packageIndices{stmgp}
}
\author{
Maintainer: \packageMaintainer{stmgp}
}
\references{
Ueki M, Tamiya G, and for Alzheimer's Disease Neuroimaging Initiative. (2016) Smooth-thresholdmultivariate genetic prediction with unbiased model selection. Genet Epidemiol 40:233-43. <https://doi.org/10.1002/gepi.21958>

Ueki M. (2009) A note on automatic variable selection using smooth-threshold estimating equations. Biometrika 96:1005-11. <https://doi.org/10.1093/biomet/asp060>
}
\keyword{ package }

